+++ src/main/java/https/github/com/wallas5h/LaskoMed/infrastructure/database/repository/jpa/DoctorAvailabilityJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoctorAvailabilityJpaRepository extends JpaRepository<DoctorAvailabilityEntity, Integer> {
+public interface DoctorAvailabilityJpaRepository extends JpaRepository<DoctorAvailabilityEntity, Long> {
 
 }
